requirements-web.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9

pydantic==2.7.3
sqlmodel==0.0.18
asyncpg==0.29.0
httpx==0.27.0
pyjwt==2.8.0
jinja2==3.1.3

aiofiles==23.2.1